# Flask/rentorlend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rentorlend API",
    "description": "API documentation for Rentorlend",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/user/{user_id}": {
      "get": {
        "summary": "Get User Details",
        "description": "Retrieve details of a specific user by user ID.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer",
                  "description": "ID of the user"
                },
                "public_id": {
                  "type": "string",
                  "description": "Public ID of the user"
                },
                "name": {
                  "type": "string",
                  "description": "Name of the user"
                },
                "email": {
                  "type": "string",
                  "description": "Email of the user"
                },
                "mobile_no": {
                  "type": "string",
                  "description": "Mobile number of the user"
                },
                "is_vendor": {
                  "type": "boolean",
                  "description": "Indicates if the user is a vendor"
                },
                "is_customer": {
                  "type": "boolean",
                  "description": "Indicates if the user is a customer"
                },
                "is_admin": {
                  "type": "boolean",
                  "description": "Indicates if the user has admin privileges"
                },
                "created_date": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Date the user was created"
                },
                "updated_date": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Date the user was last updated"
                },
                "user_image": {
                  "type": "string",
                  "description": "URL of the user's profile image"
                },
                "latitude": {
                  "type": "number",
                  "description": "Latitude of the user's location"
                },
                "longitude": {
                  "type": "number",
                  "description": "Longitude of the user's location"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          },
          "500": {
            "description": "Failed to retrieve user details",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Failed to retrieve user details"
                },
                "error": {
                  "type": "string",
                  "example": "Error details"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticate a user and retrieve a token",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}